requirements.txt
tornado==4.0.2
swirl==0.1.2
Pillow==2.7.0
